services/spider/requirements.txt
alembic==1.4.3
amqp==2.6.1
appdirs==1.4.4
billiard==3.6.3.0
celery==4.3.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
configobj==5.0.6
Cython==0.29.21
Flask==1.1.2
Flask-Migrate==2.4.0
Flask-Redis==0.3.0
Flask-SQLAlchemy==2.3.2
gevent==1.4.0
greenlet==0.4.15
grpcio==1.33.2
grpcio-tools==1.33.2
gunicorn==19.9.0
idna==2.8
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.11
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==2.19.2
protobuf==3.14.0
psutil==5.7.3
psycopg2-binary==2.7.6.1
pyee==7.0.4
pyppeteer==0.2.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.4
raven==6.10.0
redis==3.5.3
requests==2.22.0
six==1.15.0
SQLAlchemy==1.3.20
tqdm==4.52.0
urllib3==1.25.11
vine==1.3.0
webargs==4.0.0
websockets==8.1
Werkzeug==1.0.1
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0
